+++ src/model/room.ts
@@ -1,6 +1,6 @@
 import { Peer } from "./peer";
 import { Router } from "mediasoup/node/lib/Router.js";
-import { UserProducerIdSet } from "./UserProducerIdSet";
+import { UserProducerIdSet } from "./user_producer_id_set";
 
 export class Room {
